Circulation

LiMa.AppCore.Circulation — Module

Lending services of the library

The core lending services are lending and returning of books. Additional functions deal with billing (of fees for books which are overdue or damaged).

LiMa.AppCore.Circulation.Lending — Type

The process of lending of one book by one library user

LiMa.AppCore.Circulation.Reservation — Type

Reservation of a title (in case that all of its copies are in circulation)

LiMa.AppCore.Circulation.checkout_book — Function

checkout_book(book::Book, user::User)

user lends book from the library

LiMa.AppCore.Circulation.return_book — Function

return_book(book::Book)

book (which must be in circulation) gets returned to the library

LiMa.AppCore.Circulation.extend_period — Function

extend_period(lending::Lending)

Extend the lending period of lending by the standard extension period of the library. May only be called, if an extension is possible.

LiMa.AppCore.Circulation.is_overdue — Function

is_overdue(lending::Lending)::Bool

Is the lending overdue?

LiMa.AppCore.Circulation.invoice_overdue_fee — Function

invoice_overdue_fee(user::User, overdue::Lending)

Send an invoice to user for the overdue lending.

LiMa.AppCore.Circulation.invoice_damage — Function

invoice_damage_fee(user::User, book::Book)

Bill user for the book he/she damaged.