TitleCatalog

LiMa.AppCore.TitleCatalog — Module

Management of titles within the libraries catalog

The catalog contains all titles the library offers for lending (via its own inventory or by partnership with other libraries).

Administration of the libraries catalog

LiMa.AppCore.TitleCatalog.catalog_title! — Function

catalog_title!(title::Title)

Makte title available in the libraries catalog.

catalog_title!(titleID::String)

Makte the Title identified by titleID available in the libraries catalog.

LiMa.AppCore.TitleCatalog.deactivate_title! — Function

deactivate_title!(title::Title)

Remove title from the libraries catalog.

deactivate_title!(title::Title)

Remove the Title identified by titleID from the libraries catalog.

LiMa.AppCore.TitleCatalog.import_catalog! — Function

import_catalog!(catalog::AbstractArray{Title})

Add the Titles given in catalog to the libraries catalog. This occurs typically when catalog data is imported from third parties.

Administration of titles and authors

LiMa.AppCore.TitleCatalog.create_title — Function

create_title(titleID::String, isbn::String, name::String, authors::AbstractArray{Author}, publisher::String, pub_date::Date)

Create a new Title with the given attributes. The new Title isn't yet part of the libraries catalog.

LiMa.AppCore.TitleCatalog.delete_title! — Function

delete_title!(title::Title)

Delete title from the list of known titles.

LiMa.AppCore.TitleCatalog.create_author — Function

create_author(name1::String, name2::String, name3::String, affiliation::String)

Create a new Author with the given attributes.

LiMa.AppCore.TitleCatalog.delete_author! — Function

delete_author!(author::Author)

Delete author from the list of known authors.