UserManager

LiMa.AppCore.UserManager — Module

Management of library users

Only registerd library users may use the libraries services. Users breaking the libraries rules may be denied further services (they will be added to a blacklist).

As each user is associated with an address, the module is also responsible for address management.

Administration of the lifecycle of library users

LiMa.AppCore.UserManager.create_user — Function

create_user(name::String, firstname::String, birthdate::Date, address::Address)::User

Create a new user with the information given. Return the new user.

LiMa.AppCore.UserManager.update_user_data! — Function

update_user_data!(userID::Integer; firstname = nothing, birthdate = nothing, address = nothing)::User

Update the user information for user userID. Return the updated user. The information passed via the optional parameters will be updated.

LiMa.AppCore.UserManager.activate_user! — Function

activate_user!(user::User)

Make user an active library user. The services of the library are only available to active users.

LiMa.AppCore.UserManager.deactivate_user! — Function

deactivate_user!(user::User)

Deactivate user, so he/she won't be able to use the libraries services any more. This may occur, when a user doesn't pay the fees in time or if he hasn't used the libraries services for more than three years.

LiMa.AppCore.UserManager.delete_user! — Function

delete_user!(userID::Integer)

Delete the user identified by userID from the list of known users.

Retrieving library users

LiMa.AppCore.UserManager.get_user — Function

get_user(userID::Integer)::User

Return the user with ID userID. The user must exist (if not, a UserNotExists exepction occurs).

LiMa.AppCore.UserManager.user_exists — Function

user_exists(userID::Integer)::Bool

Check, if a user with userID exists.

LiMa.AppCore.UserManager.search_users — Function

search_users(firstname = nothing, birthdate = nothing, address = nothing)::AbstractVector{User}

Search for users matching the attributes given. If several attributes are given, all must match. If no matching users are found, an empty list is returned.