StorageManager

LiMa.StorageManager — Module

A persistent data pool

Overview

The StorageManager is an object pool which stores its objects permanently.

A pool is basically a container for objects. It manages these objects over their lifetime and offers some additional services for them, e.g. access control, sharing across systems or, as in our case, persistence.

A pool offers the following interface:

insert!(pool,obj) -> id: Insert an object (that has been created outside) into the pool and return its id

create!(pool,type,attrs) -> obj: Create (within the pool) a new object of type using the attributes attrs and return it

update!(pool,type,id,attrs) –> obj: Update the object objof type with id with the new attribute attrsand return the updated object

exists(pool,type,id) -> Bool: Does an object obj of type with id exist within the pool?

get(pool,type,id) -> obj: Retrieve an object obj of type with id from the pool (exactly one)

delete!(pool,type,id): Delete an exisiting object obj of type with id from the pool

lookup(pool,query,args) -> Iterable<obj>: Search for objects using a query with search terms; the result is an iterable collection of objects (which may contain 0, 1 oder more elements)

This interface definition differs from some more generic pool interfaces as it assumes that each object is of a specific type and has some unique id. We come to that point in more detail lateron.

The StorageManager offers persistence with transaction control. Therefore it has also the following transaction interface:

begin_transaction(pool): Mark the begin of a transaction

commit(pool): End a transaction and make all changes permanent

rollback(pool): End a transaction and rollback all changes

set_autocommit(pool, Bool): Sets the autocommit-mode which is by default true. In this case each operation is considered being a single transaction which is commited immediately after execution. The above mentioned transaction commands are only meaningful, if this mode is set to false.

Data managed by StorageManager

Requirements

The data managed by the StorageManager has to fullfill certain requirements:

it must be a used-defined mutable composite type (UDMCT)

it must have an attribute id::Int64 which is managed by the StorageManager; this id is a unique identifier for all objects of a specific type

its attributes may have one of the following data types

a Julia primitive type, a STRING or a DATE (i.e. types which can be easily mapped to a corresponding SQL data type)

another UDMCT

an iterable collection, containing objects of the same type (i.e. all objects returned by the Iterator for this collection have an identical eltype )

Examples

The following definitions of UDMCTs from the LiMa-Application are typical examples of data which can be managed by StorageManager:

mutable struct User 
    id::Int64            # managed by `StorageManager`
    name::String
    firstname::String
    birthdate::Date
    address::Address      # another UDMCT-type
end

mutable struct Address 
    id::Int64
    street::String
    housenumber::String
    city::String
    zip::String
    country::String
end

Most of their attributes are „simple“ data types (category 1); only User declares one attribute (address) whose type is an UDMCT.

The next definitions contain an example, where a collection-type is used:

mutable struct Author
    id::Int64
    name1::String       # firstname
    name2::String       # more names (if they exist)
    name3::String       # familyname, organisation (if not a person)
    affiliation::String
end

mutable struct Title 
    id::Int64
    titleID::String     # unique identifier used by the library
    isbn::String        # not all titles may hav a ISBN
    name::String        # title
    authors::AbstractArray{Author}
    publisher::String
    pub_date::Date
end

A Title may have one or more authors. Therefore the attribute author is defined as an AbstractArray{Author}, which is an iterable collection.

Use of StructTypes

TODO

StorageManager – Public Interface

Usage of the storage pool

LiMa.StorageManager.insert! — Function

insert!(pool::StoragePool, obj::Any)

Insert an object (created outside) into the storage pool.

LiMa.StorageManager.create! — Function

create!(pool::StoragePool, type::DataType, attrs::Tuple) --> obj

Create an object of type type from attrs within the pool and return that new object.

LiMa.StorageManager.update! — Function

update!(pool::StoragePool, type::DataType, id::Int64, attrs::Tuple) --> obj

Update the existing object of type type identified by id with the new attributes attrs and return the updated object.

LiMa.StorageManager.exists — Function

exists(pool::StoragePool, type::DataType, id::Int64) --> Bool

Is there an object of type type identified by id within the pool?

LiMa.StorageManager.get — Function

get(pool::StoragePool, type::DataType, id::Int64) --> obj

Retrieve an object of type type identified by id.

LiMa.StorageManager.delete! — Function

delete!(pool::StoragePool, type::DataType, id::Int64)

Delete the object of type type identified by id from the pool.

LiMa.StorageManager.lookup — Function

lookup(pool::StoragePool, query::String, args::Tuple) --> Iterable{obj}

Search for objects using a query with search terms; the result is an iterable collection of objects (which may contain 0, 1 oder more elements)

Schema Administration

LiMa.StorageManager.add_type! — Function

add_type!(pool::StoragePool, type::DataType)

Add type to the data scheme.

LiMa.StorageManager.delete_type! — Function

delete_type!(pool::StoragePool, type::DataType)

Delete type from the data scheme. This is only possible, if no objects of that type exist.

LiMa.StorageManager.declare_attr_unique — Function

declare_attr_unique(pool::StoragePool, type::DataType, attr_name::Symbol)

The attribute attr_name of type is unique.

LiMa.StorageManager.declare_obj_unique — Function

declare_obj_unique(pool::StoragePool, type::DataType)

Each object of type type is unique.

Transaction Management

LiMa.StorageManager.begin_transaction — Function

begin_transaction(pool::StoragePool)

Mark the begin of a transaction.

LiMa.StorageManager.commit — Function

commit(pool::StoragePool)

End a transaction and make all changes permanent.

LiMa.StorageManager.rollback — Function

rollback(pool::StoragePool)

End a transaction and rollback all changes.

LiMa.StorageManager.set_autocommit — Function

set_autocommit(pool::StoragePool, mode::Bool)

Sets the autocommit-mode which is by default true. In this case each operation is considered being a single transaction which is commited immediately after execution. The above mentioned transaction commands are only meaningful, if this mode is set to false.

Persistence Layer

Object Creation

In Julia a generic creation mechanism is possible using meta programming. E.g. the following statement

e = Expr(:call, Address, 1, "a1", "a2", "a3", "a4", "a5")

creates a valid Julia expression (in this case a constructor for a new Address), which can be evaluated using eval, thus creating an Address- object.

This can be used to implement the create-function.