UserManager

The first part of the application core to be implemented is the UserManager.

LiMa.AppCore.UserManager.create_user — Method

create_user(name::String, firstname::String, birthdate::Date, address::Address)::User

Create a new user with the information given. Return the new user.

LiMa.AppCore.UserManager.delete_user — Method

delete_user(userID::Integer)::Bool

Delete the user userID from the system. Return true if such a user existed and could be deleted.

LiMa.AppCore.UserManager.get_user — Method

get_user(userID::Integer)::User

Return the user with ID userID. The user must exist (if not, a UserNotExists exepction occurs).

LiMa.AppCore.UserManager.search_users — Function

search_users(firstname = nothing, birthdate = nothing, address = nothing)::AbstractVector{User}

Search for users matching the attributes given. If no matching users are found, an empty list is returned.

LiMa.AppCore.UserManager.update_user — Method

update_user(userID::Integer; firstname = nothing, birthdate = nothing, address = nothing)::User

Update the user information for user userID. Return the updated user. The information passed via the optional parameters will be updated.

LiMa.AppCore.UserManager.user_exists — Method

user_exists(userID::Integer)::Bool

Check, if a user with userID exists.